@@ -1,7 +1,6 @@
 import {Navigate} from 'react-router-dom';
 //Redirect
 import {useSelector} from 'react-redux';
-import React from 'react';
 
 
 const RouteGuard = ({children, ...route}) => {
@@ -13,16 +12,16 @@ const RouteGuard = ({children, ...route}) => {
   // console.log(route,token);
 
   // 判断是否登录~~~
-  // if (route.auth && !token) return (<Navigate to="/welcome" replace/>);
+  if (route.auth && !token) return (<Navigate to="/login" replace/>);
   // 如果已经登录
   if (route.path === '/login' && token) {
     //return <Redirect to="/"/>;
     return <Navigate to="/" replace/>;
   }
-  if (route.path === '/welcome' && token) {
-    // return <Redirect to="/"/>;
-    return <Navigate to="/" replace/>;
-  }
+  // if (route.path === '/welcome' && token) {
+  //   // return <Redirect to="/"/>;
+  //   return <Navigate to="/" replace/>;
+  // }
   return children;
 };
 
