@@ -3,10 +3,10 @@ import backend from 'i18next-http-backend';
 
 import {initReactI18next} from 'react-i18next';
 
-// const jlrLocal = localStorage.getItem('jlr');
-// const parseLocal = jlrLocal?JSON.parse(jlrLocal):"{}";
-// const parseSettings = parseLocal.settings? JSON.parse(parseLocal.settings):{language:'en-US'};
-// const localLng = parseSettings.language;
+const desktronLocal = localStorage.getItem('desktron');
+const parseLocal = desktronLocal ? JSON.parse(desktronLocal) : "{}";
+const parseSettings = parseLocal.settings ? JSON.parse(parseLocal.settings) : {language: 'zh-CN'};
+const localLng = parseSettings.language;
 /*
 * .init({
         react: {
@@ -40,9 +40,26 @@ import {initReactI18next} from 'react-i18next';
     });*/
 
 i18n.use(backend).use(initReactI18next).init({
+  compatibilityJSON: 'v3',
   react: {
     // 是否需要在最外层加入Suspense标签
     useSuspense: false
+  },
+  backend: {
+    // path where resources get loaded from, or a function
+    // returning a path:
+    // function(lngs, namespaces) { return customPath; }
+    // the returned path will interpolate lng, ns if provided like giving a static path
+    // the function might return a promise
+    //
+    // If allowMultiLoading is false, lngs and namespaces will have only one element each,
+    // If allowMultiLoading is true, lngs and namespaces can have multiple elements
+    loadPath: './locales/{{lng}}/{{ns}}.json',
+
+    // path to post missing resources, or a function  
+    // function(lng, namespace) { return customPath; }
+    // the returned path will interpolate lng, ns if provided like giving a static path
+    addPath: './locales/add/{{lng}}/{{ns}}',
   },
   initImmediate: true,
   languages: ['en-US', 'zh-CN'],
@@ -60,10 +77,10 @@ i18n.use(backend).use(initReactI18next).init({
   // lng: localLng,
   //选择默认语言，选择内容为上述配置中的key，即en-US/zh-CN
   // fallbackLng: "zh-CN", "en-US"
-  fallbackLng: 'zh-CN', // textLanguage.value,
+  fallbackLng: localLng, // textLanguage.value,
   defaultNS: ['translation'],
   fallbackNS: false,
-  lng: 'zh-CN',
+  lng: localLng,
   ns: 'translation',
   // keySeparator:true,
   interpolation: {
