@@ -1,22 +1,10 @@
 export const USERINFO = {
   UPDATEUSERINFO: 'UPDATEUSERINFO',
-  GATEUSERINFO: 'GATEUSERINFO',
-  UPDATEUSERTOKEN: 'UPDATEUSERTOKEN'
-};
-
-export const TESTUSER = {
-  GETUSERINFO: 'GETUSERINFO'
+  GTEUSERINFO: 'GTEUSERINFO',
 };
 
 export const SETTINGS = {
   UPDATELANGUAGE: 'UPDATELANGUAGE',
-  UPDATETHEME: 'UPDATETHEME',
-  UPDATESCREEN: 'UPDATESCREEN',
-  UPDATESETTINGS: 'UPDATESETTINGS',
-  UPDATENETWORKCONNECTIVITY: 'UPDATENETWORKCONNECTIVITY',
-  UPDATENETISCONNECTED: 'UPDATENETISCONNECTED',
-  UPDATEHAPTICFEEDBACK: 'UPDATEHAPTICFEEDBACK',
-  UPDATECCFFNTUPPERDISPHAPTIC:'UPDATECCFFNTUPPERDISPHAPTIC'
-
+  CHANGELANGUAGE: 'CHANGELANGUAGE'
 };
 
