@@ -1,4 +1,5 @@
-import {createStore, applyMiddleware} from 'redux';
+// import {createStore,applyMiddleware} from 'redux';
+import { configureStore } from '@reduxjs/toolkit'
 import createSagaMiddleware from 'redux-saga';
 import {persistStore, persistReducer} from 'redux-persist';
 import rootReducer from './reducers';
@@ -7,16 +8,19 @@ import storage from 'redux-persist/lib/storage';
 
 const persistConfig = {
   keyPrefix: '',
-  key: 'zuler',
+  key: 'desktron',
   storage,
   blacklist: ['_persist'],
-  whitelist: []  // ['settings', 'userInfo']
+  whitelist: ['userInfo','settings']  // ['settings', 'userInfo']
 };
 
 const persistedReducer = persistReducer(persistConfig, rootReducer);
 const sagaMiddleware = createSagaMiddleware();
 
-
-export const store = createStore(persistedReducer, applyMiddleware(sagaMiddleware));
+// export const store = createStore(persistedReducer,applyMiddleware(sagaMiddleware));
+export const store = configureStore({
+  reducer:persistedReducer,
+  middleware:(getDefaultMiddleware)=>(getDefaultMiddleware({serializableCheck:false}).concat(sagaMiddleware))
+});
 sagaMiddleware.run(rootSagas);
 export const persistor = persistStore(store);
