@@ -1,7 +1,8 @@
 import {all} from 'redux-saga/effects';
-import testuserSaga from './testuser'
+import userSaga from './user';
+import settingsSaga from './settings';
 
 const rootSagas = function *() {
-    yield all([testuserSaga()]);
+    yield all([settingsSaga(),userSaga()]);
 };
 export default rootSagas;
